+++ emlx/c_src/emlx_fast/qwen3.hpp
